+++ R/utils-plumber.R
@@ -59,27 +59,27 @@ check_param_num <- function(value, valid_values) {
 #' @return logical
 #' @noRd
 check_param_lgl <- function(value) {
-  out <- is.logical(value)
+  out <- is.logical(value) & !is.na(value)
 
   return(out)
 }
 
 #' Format error
 #' Format error if check_parameters() returns TRUE
-#' @param param character or numeric: Parsed parameters
-#' @param valid_values character or numeric: Accepted values for this parameter
+#' @param params character: Vector with parsed parameters
+#' @param query_controls list: Query controls
 #' @return list
 #' @noRd
-format_error <- function(param, valid_values) {
-  out <- list(
-    error = paste0(
-      "Invalid value for ", param, ". Please use one of",
-      paste(
-        "'", valid_values, "'",
-        sep = "", collapse = ", "
-      ), "."
-    )
-  )
+format_error <- function(params, query_controls) {
+  msg1 <- "You supplied an invalid value for %s. Please use one of the valid values."
+  msg2 <- "Invalid query arguments have been submitted."
+  # params <- names(params)
+  out <- lapply(params, function(x) {
+    list(msg = sprintf(msg1, x),
+         valid = query_controls[[x]]$values)
+  })
+  names(out) <- params
+  out <- list(error = msg2, details = out)
   return(out)
 }
 
@@ -101,6 +101,7 @@ validate_query_parameters <-
     "ppp",
     "version",
     "format",
+    "table",
     "parameter"
   )) {
     params$argsQuery <-
@@ -175,4 +176,43 @@ serializer_switch <- function() {
       }
     )
   }
+}
+
+#' Assign PIP API required parameters if missing
+#'
+#' @param req list: plumber `req` object
+#'
+#' @return list
+#' @noRd
+assign_required_params <- function(req) {
+
+  # Handle required names for /pip endpoint
+  endpoint <- extract_endpoint(req$PATH_INFO)
+  if (endpoint %in% c("pip", "pip-grp")) {
+    if (is.null(req$args$country)) {
+      req$args$country <- "all"
+      req$argsQuery$country <- "all"
+    }
+    if (is.null(req$args$year)) {
+      req$args$year <- "all"
+      req$argsQuery$year <- "all"
+    }
+    if (endpoint == "pip-grp") {
+      if (is.null(req$args$group_by)) {
+        req$args$group_by <- "none"
+        req$argsQuery$group_by <- "none"
+      }
+    }
+  }
+  return(req)
+}
+
+#' helper function to extract endpoint from req$PATH_INFO object
+#'
+#' @param path character: Information returned by req$PATH_INFO
+#' @return character
+#' @noRd
+extract_endpoint <- function(path) {
+  # stringr::str_extract(path, pattern = "([^/]+$)")
+  sub(".*[/]", "", path)
 }