+++ tests/testthat/test-ui_functions.R
@@ -3,14 +3,14 @@ skip_if(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER") == "")
 
 # constants
 lkups <- create_versioned_lkups(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER"))
-lkups <- lkups$versions_paths$latest_release
+lkups <-lkups$versions_paths[[lkups$latest_release]]
 set.seed(42)
 lkups$pl_lkup <- lkups$pl_lkup[sample(nrow(lkups$pl_lkup), 10)]
 lkups2 <- lkups
 lkups2$svy_lkup <- lkups2$svy_lkup[country_code %in% c('AGO', 'ZWE')]
 lkups2$ref_lkup <- lkups2$ref_lkup[country_code %in% c('AGO', 'ZWE')]
-dt_lac <- readRDS("../testdata/pip_lac_resp.RDS")
-dt_sas <- readRDS("../testdata/pip_sas_resp.RDS")
+dt_lac <- readRDS("../testdata/pip_lac_resp.rds")
+dt_sas <- readRDS("../testdata/pip_sas_resp.rds")
 
 test_that("ui_hp_stacked() works as expected", {
   res <- ui_hp_stacked(povline = 1.9, lkup = lkups2)
@@ -45,7 +45,7 @@ test_that("ui_pc_charts() works as expected", {
   # Regular query (fill_gaps = FALSE)
   res <- ui_pc_charts(country = "AGO", povline = 1.9, lkup = lkups)
   expect_equal(nrow(res), nrow(lkups$svy_lkup[country_code == "AGO"]))
-  expect_equal(length(names(res)), 37)
+  expect_equal(length(names(res)), 35)
 
   # Regular query (fill_gaps = TRUE)
   res <- ui_pc_charts(country = "AGO", povline = 1.9, fill_gaps = TRUE, lkup = lkups)
@@ -65,7 +65,7 @@ test_that("ui_pc_regional() works as expected", {
   expect_identical(
     names(res),
     c(
-      "region_code", "reporting_year",
+      "region_name", "region_code", "reporting_year",
       "reporting_pop",  "poverty_line",
       "headcount", "poverty_gap",
       "poverty_severity", "watts",
@@ -99,7 +99,7 @@ test_that("ui_cp_poverty_charts() works as expected", {
   expect_equal(nrow(dl$pov_trend),
                nrow(lkups$svy_lkup[country_code == "ARG"]))
   expect_equal(nrow(dl$pov_mrv), 11)
-
+  expect_equal(unique(dl$pov_trend$reporting_level), "urban")
   dl <- ui_cp_poverty_charts(
     country = "SUR",
     povline = 1.9,
@@ -123,9 +123,21 @@ test_that("ui_cp_poverty_charts() works as expected", {
   expect_equal(nrow(dl$pov_trend),
                nrow(lkups$dist_stats[country_code == "CHN" &
                           reporting_level == "national"]))
-  skip("These tests can be activated if we implement pipapi#149")
   expect_equal(unique(dl$pov_trend$reporting_level), "national")
-  expect_equal(unique(dl$pov_mrv$reporting_level), "national")
+
+  # Test that ui_cp_poverty_charts() works correctly for
+  # countries w/ multiple reporting levels (only national rows are returned)
+  dl <- ui_cp_poverty_charts(
+    country = "URY",
+    povline = 1.9,
+    pop_units = 1e6,
+    lkup = lkups,
+    pov_lkup = NULL
+  )
+  expect_equal(nrow(dl$pov_trend),
+               nrow(lkups$dist_stats[country_code == "URY" &
+                                       reporting_level == "national"]))
+  expect_equal(unique(dl$pov_trend$reporting_level), "national")
 })
 
 test_that("cp_pov_mrv_select_values() works as expected", {
@@ -223,6 +235,17 @@ test_that("ui_cp_ki_headcount() works as expected", {
   # aggregated distributions (only national rows are returned)
   df <- ui_cp_ki_headcount(country = "CHN", povline = 1.9, lkup = lkups)
   expect_false(is.na(df$headcount))
+  expect_equal(df$reporting_year,
+               max(lkups$dist_stats[country_code == "CHN" &
+                                    reporting_level == "national"]$reporting_year))
+  # Test that ui_cp_ki_headcount() works correctly for countries
+  # w/ multiple reporting levels (only national rows are returned)
+  df <- ui_cp_ki_headcount(country = "URY", povline = 1.9, lkup = lkups)
+  expect_false(is.na(df$headcount))
+  expect_equal(df$reporting_year,
+               max(lkups$dist_stats[country_code == "URY" &
+                                    reporting_level == "national"]$reporting_year))
+
 })
 
 test_that("ui_cp_key_indicators() works as expected", {
@@ -253,6 +276,20 @@ test_that("ui_cp_key_indicators() works as expected", {
     )
   )
   expect_identical(dl[[2]]$headcount$poverty_line, 1.9)
+
+  # All countries and poverty lines
+  dl <- ui_cp_key_indicators(country = "all", lkup = lkups2)
+  expect_length(dl, 2) # 2 countries
+  expect_length(dl[[1]], 7) # # 7 KI objects
+  expect_equal(nrow(dl[[1]]$headcount), 10) # 10 poverty lines
+  expect_identical(
+    names(dl[[2]]),
+    c(
+      "headcount", "headcount_national", "mpm_headcount",
+      "reporting_pop", "gni", "gdp_growth", "shared_prosperity"
+    )
+  )
+  expect_identical(nrow(dl[[2]]$headcount), length(lkups2$pl_lkup$name))
 
   # Only CP relevant surveys
   dl <- ui_cp_key_indicators(country = "POL", povline = 1.9, lkup = lkups)